Moved type definition into seperate file
