+++ src/types/utilTypes.ts
@@ -1,0 +1,8 @@
+export interface clientDimensions {
+  width: number;
+  height: number;
+}
+export interface DragOffset {
+  offsetX: number;
+  offsetY: number;
+}
